Added recognition of global data that needs to be wrapped around
Sample case:

 - GFS global forecast spans 0 - 359.75
 - Grid size is 1440
 - (359.75-0)*1441/1440 = 360 implying the data covers the globe

Hence we also contour the grid cell starting from the last X
coordinate and ending with the first X coordinate to completely cover
the globe (359.75 - 0). The test is done with an epsilon of 1e-3,
since floating point calculations are not accurate. In the case
of GFS the error is roughly 1e-4.