World wraparound detection moved to Q objects.
The detection of global data that needs an extra column to fully cover
the world needs to be done by the caller. For example, if we contour
the global GFS data in WebMercator, the engine previously did not see
the geographic coordinates and hence could not detect that wraparound
was needed.