Added protection agains server termination by catching contouring exceptions
Should contouring throw an exception, the thread doing the conturing
will now catch it, print a stack trace, and continue execution. This in
effect means the user gets only partial output. However, the alternative
would be no output at all, so this solution is a compromise taking
into account the fact that exceptions are known to be very rare.